infra/requirements.txt
aws-cdk-lib>=2.130.0,<3
constructs>=10.0.0,<11
